homeRoutes

+++ controllers/api/postRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { Post, User, Note } = require('../models')
+const { Post, User, Note } = require('../../models')
 const passport = require('passport')
 
 // get all posts
@@ -9,24 +9,38 @@ router.get('/posts', passport.authenticate('jwt'), async function (req, res) {
 })
 
 //get one post 
-router.get('/posts/:id', passport.authenticate('jwt'), async function (req, res) {
+router.get('/:id', passport.authenticate('jwt'), async function (req, res) {
   const posts = await Post.findOne({ where: {id: req.params.id}, include: [User, Note] })
   res.json(posts)
 })
 
 //post one post
-router.post('/posts', passport.authenticate('jwt'), async function(req, res) {
+router.post('/', passport.authenticate('jwt'), async function(req, res) {
   const post = await Post.create({
     body: req.body.body,
     title: req.body.title,
-    uid: req.user.id
+    uid: req.session.userId
   })
   res.json(post)
 })
 
-//delete a post
-router.delete('/posts/:id', passport.authenticate('jwt'), async function (req, res) {
-  await Post.destroy({ where: {id: req.params.id } })
+///edit a ppost n shit
+router.put('/:id', passport.authenticate('jwt'), async function(req, res) {
+  const post = await Post.update({
+    body: req.body.body,
+    title: req.body.title},
+    {where: {id:req.params.id}}
+    res.json(post)
+  })
+})
+
+//delete a post or comment
+router.delete('/:id', passport.authenticate('jwt'), async function (req, res) {
+  let post = await Post.destroy({ where: {id: req.params.id, uid: req.session.userId } })
+
+  if(post > 0) {
+    let notes = await Note.destroy({where : {uid: req.params.id}})
+  }
   res.sendStatus(200)
 })
